requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
qdrant-client==1.7.0
langgraph==0.0.49
crewai==0.28.8
python-dotenv==1.0.0
openai==1.13.3
python-telegram-bot==20.7
twilio==8.11.0
requests==2.31.0
websockets==12.0
